pyproject.toml
[project]
name = "robotframework-oneflow"
version = "0.1.0"
description = "Robot Framework library to treat a test suite as a one flow, where if any test case will fail than it skip the rest of the tests"
authors = [{ name = "otemek", email = "etombsc@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["robotframework (>=6)"]

[tool.poetry]
packages = [{ include = "OneFlow" }]


[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
